+++ CHANGELOG.md
@@ -8,6 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line.
 
+### Added
+
+- Notify: Add BCC and CC fields to SendEmail notification action.
+- Add the CHANGELOG.md to the root of the code base.
+
+### Changed
+
+- Notify: Changed the Email topology type to support comma-separated list of emails when using constants.
+
 ## Older versions
 
 Find older release notes [here](./doc/changelog.rst).